Move Transformations and Filters in AppGateway to subfolders for better structure
